Lint app/ directory with rubocop

+++ app/workers/expire_file_cache.rb
@@ -1,7 +1,8 @@
+# frozen_string_literal: true
 require 'find'
 class ExpireFileCache
   include Sidekiq::Worker
-  sidekiq_options :queue => :file_cache
+  sidekiq_options queue: :file_cache
 
   def self.display_name
     'Expiring cache entries'
@@ -15,9 +16,7 @@ class ExpireFileCache
       Find.find(Rails.cache.cache_path) do |path|
         if FileTest.file?(path)
           c = Marshal.load(File.read(path))
-          if c.expired?
-            File.unlink(path)
-          end
+          File.unlink(path) if c.expired?
         end
       end
     end
@@ -25,9 +24,8 @@ class ExpireFileCache
 
   def other_expirers_running?
     workers = Sidekiq::Workers.new
-    workers.any? do |process_id, thread_id, work|
-      work['payload']['jid'] != self.jid && work['queue'] == 'file_cache'
+    workers.any? do |_process_id, _thread_id, work|
+      work['payload']['jid'] != jid && work['queue'] == 'file_cache'
     end
   end
-
 end
